requirements.txt
absl-py==1.0.0
numpy==1.21.5
cloudpickle==2.0.0
six==1.16.0
dm-acme==0.4.0
libpython==0.2
dm-acme[tf]
chex==0.1.3
Cython==0.29.28
flax==0.4.1
optax==0.1.2
rlax==0.1.2
pyglet==1.5.24
jax==0.3.7
jaxlib==0.3.7
dm-haiku==0.0.5
dm-acme[reverb]
gym[accept-rom-license, atari, Box2D]==0.21.0
xmanager==0.1.5
ale_py==0.7.5
pyvirtualdisplay==3.0
lazydict==1.0.0b2
sk-video==1.1.10
ffmpeg-python==0.2.0
